Rename okhttp3.internal.sse to okhttp3.sse.internal
I think this is slightly neater from an OSGi/JPMS perspective.
