+++ okhttp-sse/src/test/java/okhttp3/sse/internal/EventSourceRecorder.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package okhttp3.internal.sse;
+package okhttp3.sse.internal;
 
 import java.io.IOException;
 import java.util.concurrent.BlockingQueue;
